+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Added
+  * Added an example for `terra-hyperlink` to demonstrate text ellipses when its content should overflow.
 * Changed
   * Fixed typos and imports in `terra-list` docs.